Fix bug when current King is in check.
The method is_check_mate_or_stale_mate raises an exception because the method to calculate the positions did it for every piece in the board, including the pieces for the player which is not the current. Since the current king is in check, when moving the oppenent piece that is attacking the king, it removes the current king, so when verifying if current king is in check, there was no king and the system panics. The solution is to calculate the positions only for the current player.